Fixed error in matrix subtraction
Identified problem with the centroid matrix being recycled, and not
therefore being accurately subtracted from the matrix X. Fixed here
by replicating the matrix to the same size of X

Added new function: matrix_centroids and unit tests for this function.
May be overkill, but need to rule out this part being a problem!